requirements.txt
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
colorama==0.4.4
Django==4.1
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
filelock==3.7.1
google-api-core==2.8.2
google-api-python-client==2.57.0
google-auth==2.10.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
httplib2==0.20.4
huggingface-hub==0.7.0
idna==3.3
numpy==1.22.4
oauthlib==3.2.0
packaging==21.3
protobuf==4.21.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.4.0
pyparsing==3.0.9
python-magic==0.4.27
pytz==2022.2.1
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
sentencepiece==0.1.96
six==1.16.0
sqlparse==0.4.2
tokenizers==0.12.1
tqdm==4.64.0
transformers==4.19.2
typing_extensions==4.2.0
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.9
django-extensions==3.2.1
django-filter==22.1
